Update database.cfg
Fixing some typos I found in the database.cfg, as well as cleaning up some of the discovery messages to flow better in English.

I tried not to touch them too much, because I didn't want to step on any toes, but if you want I can go through and try to come up with funny/interesting things to say for the bodies, based on the originals. (Or come up with new negative text results. Or whatever.)

Aside, sorry for forking the ENTIRE repository just to make changes to the cfg file. I'm still learning how to use git properly.